feat: support `conventional-changelog-*` presets in `conventional-recommended-bump`
Support the ability to pass in `conventional-changelog-*` preset pakage
names using the `preset` option.

Previous behavior of the `preset` option was to map the name provided
through the option to one of the presets built into this
`conventional-recommended-bump` package.

Now we take the name passed to the `preset` option, whether passed programmatically
or via the CLI, append it to `conventional-changelog-`, and then attempt to
import that package.

The loading behavior matches the behavior of the `preset` option in
`conventional-changelog`.

Furthermore, the loading behavior has been extended to support the
different export types of conventional preset packages, such as a
promise, function, or object.

Lastly, we defer to `recommendedBumpOpts.parserOpts` property of a preset
package if it exists, as our initial refactor efforts created a
`parserOpts` object under `recommendedBumpOpts` in each preset package.

In the future we want to merge differences found in
`recommendedBumpOpts.parserOpts` into the top-level `parserOpts` object
and remove `recommendedBumpOpts.parserOpts` from each presetpackage if
it exists.

BREAKING CHANGE:

First:

Will only accept the name associated with a `conventional-changelog-*`
preset package for the purpose of loading preset configuration.

Built-in presets previously available in `conventional-recommended-bump`
are no longer available.

Related to #241

Second:

Remove backwards-compatible support for `whatBump` to return a number.

If `whatBump` returns a number it will no longer be mapped to an object
with a `level` and `releaseType` property. Please update the passed in
`whatBump` function to return an object with a `level` property
indicating the type of bump (`0` for `major, `1` for `minor`, and `2`
for `patch`).

Third:

A callback function must be provided to `conventional-recommended-bump`,
either in the second argument spot, in place of `parserOpts`, or in the
third argument spot, technically reserved for a callback.